profhome_it_bots/requirements.txt
pyyaml==6.0.1

# TeleBot
pyTelegramBotAPI==4.12.0

# FastAPI
fastapi==0.101.0
pydantic==2.1.1
uvicorn==0.23.2
databases[postgresql]==0.7.0
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.11.2